+++ modules/hanna-css/src/lib/style-server-info.ts
@@ -76,6 +76,7 @@ export type CssModuleToken =
   | 'NewsHero'
   | 'PageFilter'
   | 'PageHeading'
+  | 'Pagination'
   | 'Picture'
   | 'ProcessOverview'
   | 'PullQuote'