+++ modules/hanna-react/CHANGELOG.md
@@ -3,9 +3,17 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Add component `Pagination`
+- `MegaPanelItem` type:
+  - feat: add prop `Content` for custom `*Panel__item` children
+  - feat: add prop `modifier`, similar to `MainMenuItem.modifier`
+- feat: Export `MegaPanelItemList` from `MainMenu`
+- feat: Expand `MainMenuItem.modifier` to accept falsy and array values
+- feat: Add props `wrapperProps` to `CarouselStepper`
 - feat: Remove defeault/implicit seen-effect styling from all
   components(except `SeenEffect`, `BgBox` and `CenterColumn`) — flag
   `startSeen` prop as deprecated.
+- fix: Export type `MultiselectOption`, deprecate `MultiSelectOption`
 
 ## 0.10.101